% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in later2.R
\name{format.event_loop}
\alias{format.event_loop}
\title{Format an event loop}
\usage{
\method{format}{event_loop}(x, ...)
}

\arguments{
\item{x}{An event loop object.}

\item{...}{Further arguments passed to or from other methods (currently
  unused).}
}

\value{
A character string.
}

\description{
Format an event loop
}

\keyword{internal}
