% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in promise.R
\name{format.promise}
\alias{format.promise}
\title{Format a promise}
\usage{
\method{format}{promise}(x, ...)
}

\arguments{
\item{x}{A promise object.}

\item{...}{Further arguments passed to or from other methods (currently
  unused).}
}

\value{
A character string.
}

\description{
Format a promise
}

\keyword{internal}
