% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in later2.R
\name{loop_empty}
\alias{loop_empty}
\title{Check if later loop is empty}
\usage{
loop_empty(loop = current_loop())
}

\arguments{
\item{loop}{A handle to an event loop.}
}

\description{
Returns true if there are currently no callbacks that are scheduled to
execute in the present or future.
}

\keyword{internal}
