% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in later2.R
\name{next_op_secs}
\alias{next_op_secs}
\title{Relative time to next scheduled operation}
\usage{
next_op_secs(loop = current_loop())
}

\arguments{
\item{loop}{A handle to an event loop.}
}

\description{
Returns the duration between now and the earliest operation that is currently
scheduled, in seconds. If the operation is in the past, the value will be
negative. If no operation is currently scheduled, the value will be `Inf`.
}

