% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in later2.R
\name{print.event_loop}
\alias{print.event_loop}
\title{Print an event loop}
\usage{
\method{print}{event_loop}(x, ...)
}

\arguments{
\item{x}{An event loop object.}

\item{...}{Further arguments passed to or from other methods (currently
  unused).}
}

\value{
`NULL`, invisibly. Called for its side effect of printing.
}

\description{
Print an event loop
}

\keyword{internal}
