% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\name{is.promise}
\alias{is.promise}
\title{Determine whether an object is a promise}
\usage{
is.promise(x)
}
\arguments{
\item{x}{An R object to test.}
}
\value{
`TRUE` if `x` is a promise object created by [promise()], `FALSE`
  otherwise.
}
\description{
Determine whether an object is a promise
}
