% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later2-package.R
\docType{package}
\name{later2-package}
\alias{later2}
\alias{later2-package}
\title{later2: Utilities for Scheduling Functions to Execute Later with Event Loops}
\description{
Executes arbitrary R or C functions some time after the current time, after the R execution stack has emptied. The functions are scheduled in an event loop. This is a derived work from the 'later' package aiming to reduce the number of dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/later2}
  \item Report bugs at \url{https://github.com/pachadotdev/later2/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
  \item Winston Chang \email{winston@posit.co} (\href{https://orcid.org/0000-0002-1576-2126}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
  \item Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder]
}

Other contributors:
\itemize{
  \item Marcus Geelnard (TinyCThread library, https://tinycthread.github.io/) [contributor, copyright holder]
  \item Evan Nemerson (TinyCThread library, https://tinycthread.github.io/) [contributor, copyright holder]
}

}
\keyword{internal}
