% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in debug.R
\name{logLevel}
\alias{logLevel}
\title{Get and set logging level}
\usage{
logLevel(level = NULL)
}

\arguments{
\item{level}{The logging level. Must be one of \code{NULL}, \code{"OFF"},
  \code{"ERROR"}, \code{"WARN"}, \code{"INFO"}, or \code{"DEBUG"}. If
  \code{NULL} (the default), then this function simply returns the current
  logging level.}
}

\value{
If \code{level=NULL}, then this returns the current logging level. If
  \code{level} is any other value, then this returns the previous logging
  level, from before it is set to the new value.
}

\description{
The logging level for later can be set to report differing levels of
information. Possible logging levels (from least to most information
reported) are: \code{"OFF"}, \code{"ERROR"}, \code{"WARN"}, \code{"INFO"}, or
\code{"DEBUG"}. The default level is \code{ERROR}.
}

\keyword{internal}
